Added Models and Controllers for each trend query, wrote the queries, and created table views

+++ FlightProject/Models/Flight2000.cs
@@ -52,7 +52,7 @@ namespace FlightProject.Models
         public int Airtime { get; set; }
 
         [Column("ARRDELAY")]
-        public int ArrivalDelay { get; set; }
+        public float ArrivalDelay { get; set; }
 
         [Column("DRPDELAY")]
         public int DepartureDelay { get; set; }
